Room toString and starting error
Changed the room to string to be more reflective of how we see the room numbers on campus with all the zeroes and such. I also changed the start of the program to set the error label when the class is not chosen so that it doesn't immediately throw an error to the console. Lastly, I ran the formatter on some of the classes to correct indentation and spacing.