ScheduleViewModel debug 2: I change all the files
First, I ran the formatter on each of the current classes, so most of the changes are just spacing and boring stuff

Second, I worked on trying to figure out why the update function in scheduleViewModel was not getting called properly. Now it calls it, but it causes a null pointer when trying to get the session from the session list. This is after the console prints out that it is added to the list