The time of the course now appears
Nothing else works though. Basically, the program was previously trying to pull the classgroup from the course of the session, but that was null. Even though we were adding courses to classgroups, the courses did not have the classgroup set within. I fixed this with a setter and then updated the method to add a course to a classgroup to use that setter.