Sessions display again
Had to catch the exceptions to get main to open. Then the checker for the availability of the teacher was backwards, so that had to be fixed.

Commented out line 157 in ScheduleGridViewController since you can't do getLayout on the list of children. The layout must be taken from a specific child