Remove a session on click
Inside the session details grid, you can now remove a session by clicking on it in the grid pane. I had to pray to the old gods, but it works.

First, it finds and removes the session based on the text in the label. Then it removes the label from the grid pane by comparing the text in the label to all other labels present.