Debug messages
Trying to figure out why the program slows down so much when the amount of sessions increases above 5. Currently, when the grid resets, it calls the sessionlist by class and date about 15 times. When a session is added, the suggest rooms is called 10 times. This is greatly slowing down the program