Squashed bug: Teacher list in select teacher overpopulates
Previously, the teacher list would keep adding another set of teachers to the list. This was because the array clear and the choice box clear were in the wrong order. The array would be cleared. Then the choicebox would remove everything that is present in the empty array: nothing. I also changed the array removal to use the clear method.