Work in progress: Danish site
Changed the home page text as well as all of the nav bars. I still need to translate the other pages including the titles and alts for images. Not sure about the carousels for classes. I still need to add two flags at the top right of the page to link to the different languages.